Added Plivo, Twilio.AspNet.Core, and Newtonsoft.Json to the Android and iOS projects Could not add Nexmo.Csharp.Client. The error is:
Could not install package 'runtime.debian.8-x64.runtime.native.System.Security.Cryptography.OpenSsl 4.3.0'. You are trying to install this package into a project that targets 'MonoAndroid,Version=v8.0', but the package does not contain any assembly references or content files that are compatible with that framework. For more information, contact the package author.

See https://forums.xamarin.com/discussion/85975/cannot-install-nuget-package-for-system-security-cryptography-algorithms-4-3-0.
This forum post says the error is due to the Nuget package not being supported in PCL.